Backend/requirements.txt
bcrypt==3.2.2
cffi==1.15.0
pycparser==2.21
PyJWT==2.4.0
